Refactor and simplify retrievals so the Identifier classes can interact directly
